Consolidate to a single tile
