+++ src/app/inventory/StoreBucket.scss
@@ -52,6 +52,7 @@
 
     .equipped-item {
       border: none;
+      padding-top: 0;
     }
   }
 }
