--- src/app/inventory/TallTile.scss
@@ -1,182 +1,0 @@
-@import '../../scss/variables.scss';
-
-.item-drag-container {
-  box-sizing: border-box;
-  width: var(--item-size);
-  height: calc(#{$full-height-badge});
-  margin: 0 var(--item-margin) var(--item-margin) 0;
-  &:hover {
-    outline: 1px solid #ddd;
-  }
-}
-.no-badge {
-  .item-drag-container {
-    height: var(--item-size);
-  }
-  .item {
-    height: var(--item-size);
-  }
-}
-
-.diamond .item-img {
-  border-color: transparent;
-}
-.item {
-  box-sizing: border-box;
-  width: var(--item-size);
-  height: calc(#{$full-height-badge});
-  margin: 0 var(--item-margin) var(--item-margin) 0;
-
-  .item-img {
-    box-sizing: border-box;
-    border-width: $item-border-width;
-  }
-
-  .overlay {
-    box-sizing: border-box;
-    top: $item-border-width;
-    left: $item-border-width;
-    border-width: 0px;
-    height: calc(var(--item-size) - #{2 * $item-border-width});
-    width: calc(var(--item-size) - #{2 * $item-border-width});
-    position: absolute;
-    background-repeat: no-repeat;
-    background-position: center;
-  }
-
-  .masterwork {
-    .item-img {
-      border-color: #eade8b;
-    }
-    .item-stat {
-      background-color: #eade8b;
-    }
-    &::after {
-      display: none;
-    }
-    .overlay {
-      background-size: calc((var(--item-size) - #{2 * $item-border-width}) * (96 / 90))
-        calc((var(--item-size) - #{2 * $item-border-width}) * (96 / 90));
-      background-image: url('../../images/masterwork.png');
-    }
-  }
-
-  .exotic.masterwork {
-    .overlay {
-      background-image: url('../../images/exotic-masterwork.png');
-    }
-  }
-
-  .capped {
-    .item-img {
-      border-color: $gold;
-    }
-    .item-stat {
-      background-color: $gold;
-    }
-  }
-
-  .fullstack {
-    .item-stat {
-      font-weight: bold;
-      color: $solar;
-    }
-  }
-
-  .item-stat {
-    top: calc(var(--item-size) - #{$item-border-width});
-    bottom: auto;
-    background-color: #ddd;
-    color: black;
-    height: calc(#{$badge-height});
-    font-size: calc(#{$badge-font-size});
-    width: 100%;
-    display: flex;
-    text-align: right;
-    box-sizing: border-box;
-    padding: 0 2px;
-    white-space: pre;
-    align-items: center;
-    justify-content: flex-end;
-    line-height: calc(#{$badge-height});
-  }
-
-  .primary-stat {
-    flex: 1;
-  }
-
-  .item-review {
-    font-size: calc(var(--item-size) / 5.5);
-    text-align: left;
-    display: flex;
-    align-items: center;
-    line-height: 1em;
-    transform: translateY(-0.5px);
-    .godroll {
-      color: #ff7722;
-    }
-    .app-icon {
-      font-size: 0.7em;
-      padding-right: 1px;
-      padding-left: 0;
-      position: static;
-      transform: translateY(0.5px);
-    }
-  }
-
-  .icons {
-    position: absolute;
-    left: $item-border-width + 2px;
-    top: calc(var(--item-size) - #{$badge-height} - #{$item-border-width});
-    display: flex;
-    flex-direction: row;
-  }
-
-  .item-tag {
-    // TODO: proportional size
-    display: block;
-    position: static;
-    width: calc(var(--item-size) / 5);
-    height: calc(var(--item-size) / 5);
-    margin-right: 2px;
-    color: #29f36a; // #5eff92;
-    filter: drop-shadow(0px 0px 2px rgba(0, 0, 0, 0.8));
-  }
-
-  .element {
-    width: calc(var(--item-size) / 6);
-    height: calc(var(--item-size) / 6);
-    transform: translateY(0.5px);
-    &.void {
-      filter: brightness(200%);
-      background-color: transparent !important;
-    }
-    &.arc {
-      filter: brightness(65%) saturate(200%);
-    }
-  }
-
-  .item-quality .app-icon {
-    filter: drop-shadow(0px 0px 1px rgba(0, 0, 0, 0.8));
-  }
-
-  .item-xp-bar {
-    background: rgba(0, 0, 0, 0.4);
-    position: absolute;
-    width: auto;
-    left: 3px;
-    right: 3px;
-    opacity: 0.9;
-    top: 3px;
-    height: calc(var(--item-size) / 9);
-    .item-xp-bar-amount {
-      height: 100%;
-      background-color: $xp;
-    }
-  }
-}
-
-.ms-edge .item .item-tag {
-  // https://github.com/DestinyItemManager/DIM/issues/3291
-  filter: none !important;
-}
